+++ src/Pattern/index.ts
@@ -1,0 +1,2 @@
+export { Pattern } from './Pattern';
+export type * from './types';